sub2/backend/requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-composite-foreignkey==1.1.0
django-cors-headers==3.2.1
django-extensions==2.2.5
django-imagekit==4.0.2
django-rest-swagger==2.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
entrypoints==0.3
html5lib==1.0.1
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
JPype1==0.7.3
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.12
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scikit-surprise==1.1.0
scipy==1.3.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
surprise==0.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
tweepy==3.8.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
wincertstore==0.2
zipp==3.1.0
